Add files via upload
Finished skill tree v1

+++ GameIntegration/Random.cpp
@@ -1,4 +1,4 @@
 #include "Random.h"
 #include <chrono>
 
-std::mt19937 Random::generator((unsigned int)std::chrono::system_clock::now().time_since_epoch().count());
+std::mt19937 Random::generator(std::chrono::system_clock::now().time_since_epoch().count());
